+++ src/services/Services.ts
@@ -574,12 +574,12 @@ export function validateScriptHash(output: string, outputFormat?: GetUtxoStatusO
  */
 export function toBinaryBaseBlockHeader(header: BaseBlockHeader): number[] {
   const writer = new Utils.Writer()
-  writer.writeUInt32BE(header.version)
+  writer.writeUInt32LE(header.version)
   writer.writeReverse(asArray(header.previousHash))
   writer.writeReverse(asArray(header.merkleRoot))
-  writer.writeUInt32BE(header.time)
-  writer.writeUInt32BE(header.bits)
-  writer.writeUInt32BE(header.nonce)
+  writer.writeUInt32LE(header.time)
+  writer.writeUInt32LE(header.bits)
+  writer.writeUInt32LE(header.nonce)
   const r = writer.toArray()
   return r
 }