+++ test/Wallet/get/getHeaderForHeight.test.ts
@@ -1,3 +1,8 @@
+import { Utils } from '@bsv/sdk'
+import {
+  blockHash,
+  deserializeBaseBlockHeader
+} from '../../../src/services/chaintracker/chaintracks/util/blockHeaderUtilities'
 import { _tu, TestWalletNoSetup } from '../../utils/TestUtilsWalletStorage'
 
 const includeTestChaintracks = false
@@ -38,6 +43,11 @@ describe('getHeaderForHeight tests', () => {
       // Query an existing valid block height
       const height = 1 // Ensure this height exists in the test database
       const result = await wallet.getHeaderForHeight({ height })
+      const headerHex = result.header
+      const headerA = Utils.toArray(headerHex, 'hex')
+      const hash = blockHash(headerA)
+      expect(hash).toBe('00000000b873e79784647a6c82962c70d228557d24a747ea4d1b8bbe878e1206')
+      const header = deserializeBaseBlockHeader(headerA)
 
       expect(result).toHaveProperty('header')
       expect(typeof result.header).toBe('string')