+++ CHANGELOG.md
@@ -4,6 +4,18 @@ This document captures the history of significant changes to the wallet-toolbox 
 The git commit history contains the details but is unable to draw
 attention to changes that materially alter behavior or extend functionality.
 
+## wallet-toolbox 1.6.30
+
+- Add txid index to proven_tx_reqs table in storage knex schema
+
+## wallet-toolbox 1.6.29
+
+- Add txid index to transactions table in storage knex schema
+
+## wallet-toolbox 1.6.28
+
+- Fix The method `Services`.`getHeaderForHeight` must serialize four byte values LE
+
 ## wallet-toolbox 1.6.27
 
 - Change internalizeAction Improve handling of atomic beefs containing transactions unknown to storage.